fix(package): improve npm scripts for better workflows & usage
- Mocked dev mode now watches changes in mocks and takes them into 
account after reload (using nodemon), this behaviour was in place before 
but we introduced a regression as we moved the json-server into a 
middleware. Previously the json-server (mock-server) was running 
standalone in a separate server where we were able to configure watch 
options.
- Added options to open the browser automatically on demand, previously 
it was opening the browser by default. This might be undesired. 
Especially when watching for changes in the server side logic since it 
would re-open the browser again and again.
- Restructured the order of npm scripts